Stop returning only the first price after c.EPOCH
If we `return` the price as soon as we find a day after the EPOCH that is less than the current day, the end result is that we'd only ever return the first price we configure that's after the epoch. This pushes the return out until we've actually finished with the for loop.